Added invoice pdf generation through PDFKit

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator");
+const fileHelper = require("../util/file");
 
 const getAddProduct = async (req, res, next) => {
   if (!req.session.isLoggedIn) {
@@ -12,7 +13,7 @@ const getAddProduct = async (req, res, next) => {
       editing: false,
       hasError: false,
       errorMessage: null,
-      validationErrors: []
+      validationErrors: [],
     });
   } catch (err) {
     console.log(err);
@@ -23,24 +24,24 @@ const postAddProduct = async (req, res, next) => {
   const { title, price, description } = req.body;
   const image = req.file;
   if (!image) {
-    return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Add Product',
-      path: '/admin/add-product',
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
       editing: false,
       hasError: true,
       product: {
         title: title,
         price: price,
-        description: description
+        description: description,
       },
-      errorMessage: 'Attached file is not an image.',
-      validationErrors: []
+      errorMessage: "Attached file is not an image.",
+      validationErrors: [],
     });
   }
-  const errors = validationResult(req)
+  const errors = validationResult(req);
 
   const imageUrl = image.path;
-  if(!errors.isEmpty()) {
+  if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/add-product",
@@ -53,7 +54,7 @@ const postAddProduct = async (req, res, next) => {
         imageUrl: imageUrl,
       },
       errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
   }
 
@@ -71,7 +72,7 @@ const postAddProduct = async (req, res, next) => {
     console.log("Error creating the product ", err);
     const error = new Error(err);
     error.httpStatusCode = 500;
-    return next(error)
+    return next(error);
   }
 };
 
@@ -96,21 +97,21 @@ const getEditProduct = async (req, res, next) => {
       product: editDetails,
       hasError: false,
       errorMessage: null,
-      validationErrors: []
+      validationErrors: [],
     });
   } catch (err) {
     const error = new Error(err);
     error.httpStatusCode = 500;
-    return next(error)
+    return next(error);
   }
 };
 
 const postEditProduct = async (req, res, next) => {
   const { title, price, description, productId } = req.body;
-  const image = req.file;
-  const errors = validationResult(req)
+  const imageUrl = req.file;
+  const errors = validationResult(req);
 
-  if(!errors.isEmpty()) {
+  if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/add-product",
@@ -120,10 +121,10 @@ const postEditProduct = async (req, res, next) => {
         title: title,
         price: price,
         description: description,
-        _id: productId
+        _id: productId,
       },
       errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
   }
 
@@ -139,8 +140,9 @@ const postEditProduct = async (req, res, next) => {
     product.title = title;
     product.description = description;
     product.price = price;
-    if(image) {
-      product.imageUrl = image.path;
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = imageUrl.path;
     }
 
     await product.save();
@@ -148,7 +150,7 @@ const postEditProduct = async (req, res, next) => {
   } catch (err) {
     const error = new Error(err);
     error.httpStatusCode = 500;
-    return next(error)
+    return next(error);
   }
 };
 
@@ -165,7 +167,7 @@ const getProducts = async (req, res, next) => {
   } catch (err) {
     const error = new Error(err);
     error.httpStatusCode = 500;
-    return next(error)
+    return next(error);
   }
 };
 
@@ -173,12 +175,25 @@ const postDeleteProduct = async (req, res, next) => {
   const { productId } = req.body;
 
   try {
+    const product = await Product.findById(productId);
+    if (!product) {
+      const error = new Error("Product Not Found!");
+      error.httpStatusCode = 404;
+      return next(error);
+    }
+    //safeguard deletion of file
+    try {
+      await fileHelper.deleteFile(product.imageUrl);
+    } catch (fileErr) {
+      console.error("file deletion failed", fileErr);
+    }
+    // delete the product
     await Product.deleteOne({ _id: productId, userId: req.user._id }); // Await the deleteById function
     res.redirect("/admin/products");
   } catch (err) {
     const error = new Error(err);
     error.httpStatusCode = 500;
-    return next(error)
+    return next(error);
   }
 };
